gan trainer basic workflow working
* blend code with monai framework source files
* add handlers: checkpoint saving, log stats
* add keys for gan loss metrics
